% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cut_sp.R
\name{clip_spdf}
\alias{clip_spdf}
\title{Cut SpatialPolygonsDataFrame}
\usage{
clip_spdf(shp, bb, byid = T)
}
\arguments{
\item{shp}{SpatialPolygonsDataFrame}

\item{bb}{Boundry box}

\item{byid}{boolean}
}
\description{
Cut a spdf to a bondry box
}

