% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_boundaries.R
\name{county_towns}
\alias{county_towns}
\title{Towns in Sweden}
\usage{
county_towns(date, format = c("sp", "df", "meta"))
}
\arguments{
\item{date}{a date, a year or a vector with date/year range}

\item{format}{format of return object, "df" for data.frame, "sp" for 
SpatialPointsDataFrame and "meta" for only meta-data.}
}
\description{
Get Swedish towns for a specified year 1634-1990.
}
