% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc.R
\name{create_block}
\alias{create_block}
\title{Create lowest common denominator id}
\usage{
create_block(x, y)
}
\arguments{
\item{x}{id 1}

\item{y}{id 2}
}
\description{
Foreach \code{x} and \code{y}, checks for matches in all \code{x}
  and all \code{y}, gets all matches + all matches newid and updates
  \code{newid}. Essentially find the lowest common demoninator.
}

