% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cut_sp.R
\name{cut_spbox}
\alias{cut_spbox}
\title{Cut SpatialPolygonsDataFrame}
\usage{
cut_spbox(big, small, dist, ...)
}
\arguments{
\item{big}{SpatialPolygonsDataFrame to be cut}

\item{small}{SpatialPolygonsDataFrame which to get bbox of}

\item{dist}{Added distanse of bbox}

\item{...}{optional arguments passed tp \code{clip_spdf}}
}
\description{
Cut a sp object to the boundary box of another sp object
}

