% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{eu_geom}
\alias{eu_geom}
\title{Administrative division of the European states, 1900-2003}
\format{
A \code{sf} Simple feature collection with 198 features
\describe{
  \item{country}{Country code}
  \item{name}{Country name}
  \item{year}{Year}
  \item{geom}{Geometry}
}
}
\source{
MPIDR [Max Planck Institute for Demographic Research] and CGG [Chair for
Geodesy and Geoinformatics, University of Rostock] 2013: MPIDR Population
History GIS Collection – Europe (partly based on © EuroGeographics for the
administrative boundaries). Rostock. 
The maps used in this publication are partly based on the following source: ©
EuroGeographics for the administrative boundaries.
}
\usage{
data(eu_geom)
}
\description{
Providing information on the administrative division of the European states 
at varying geographic detail in 30-years intervals (1900, 1930, 1960, 1990, 2003). 
Until 1960, geocodes are based on the IDs used in the Princeton European Fertility 
Project. The shapefile-series is partly based on a shapefile by 
EuroGeographics (see citation information)

Projection ETRS89-extended / LAEA Europe EPSG:3035
}
