% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{geom_borders}
\alias{geom_borders}
\title{Administrative borders of Sweden, 1600-1990}
\format{
A \code{sf} Simple feature collection with 9716 features
\describe{
  \item{geom_id}{Unique id}
  \item{borders}{Number of borders in collection}
  \item{ref_code}{Geocodes from the Swedish National Archive}
  \item{start}{Start year}
  \item{end}{End year}
  \item{type_id}{Unit type id}
  \item{geometry}{Geometry}
}
}
\source{
Based upon "Historiska GIS-kartor (information om territoriella indelningar i Sverige från 
1600-talets slut till 1900-talets slut)" historical GIS data from the Swedish 
National Archive released under Creative Commons CCZero.
}
\usage{
data(geom_borders)
}
\description{
Providing supplementary information on the boundaries (without the area) administrative geographical division of Swedish administrative units of:
Municipal, Pastorship, Parish, Bailiwick, Contract, Magistrates court, Hundred, District court, 
County, Diocese and Court.

Based upon historical GIS data from the Swedish National Archive. Data on parishes and counties has 
been verified, while the other types of units have not. Thus, be aware of possible inconsistencies and
faults. 

Projection SWEREF99 EPSG:3006
}
