% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{geom_relations}
\alias{geom_relations}
\title{Geographical relatives of units}
\format{
A \code{tibble} with 19958 rows and 5 columns
\describe{
  \item{g1}{geom_id of unit 1}
  \item{g2}{geom_id of unit 2}
  \item{relation}{Type of relation "pre" or "succ"}
  \item{year}{Year of relationship}
  \item{type_id}{Unit type id}
}
}
\usage{
data(geom_relations)
}
\description{
Relations between units, succeeding and preceding units.
}
