% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{geom_sp}
\alias{geom_sp}
\title{Administrative boundaries of Sweden, 1600-1990}
\format{
A sf Simple feature collection with 12433 features
\describe{
  \item{geom_id}{Unique id}
  \item{topo_id}{Unique topographic ID}
  \item{ref_code}{Geocodes from the Swedish National Archive}
  \item{name}{Unity name}
  \item{type}{Name of unit type}   
  \item{type_id}{Unit type id}
  \item{start}{Start year}
  \item{end}{End year}
  \item{geometry}{Geometry}
}
}
\source{
"Historiska GIS-kartor (information om territoriella indelningar i Sverige från 
1600-talets slut till 1900-talets slut)" historical GIS data from the Swedish 
National Archive released under Creative Commons CCZero.
}
\usage{
data(geom_sp)
}
\description{
Providing information on administrative geographical division of Swedish administrative units of:
Municipal, Pastorship, Parish, Bailiwick, Contract, Magistrates court, Hundred, District court, 
County, Diocese and Court.

Based upon historical GIS data from the Swedish National Archive. Data on parishes and counties has 
been verified, while the other types of units have not. Thus, be aware of possible inconsistencies and
faults. 

Projection SWEREF99 EPSG:3006
}
