% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hist_boundaries.R
\name{get_year}
\alias{get_year}
\title{Get year}
\usage{
get_year(x, ...)
}
\arguments{
\item{x}{An oabject to be converted}

\item{...}{other paramaters for methods}
}
\description{
Transforms a atomic vector to a integer of year
}

