% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{hist_parish}
\alias{hist_parish}
\title{Historical parish boundaries}
\format{A \code{SpatialPolygonsDataFrame} with 101 elements
\describe{
  \item{geomid}{Geomid}
  \item{nadkod}{Unique parish code}
  \item{grkod}{A code}
  \item{socken}{Parish name}
  \item{from}{Starting year}
  \item{tom}{End year}
  \item{area}{Area in km2}
  \item{dedikscb}{Dediksvb code}
  \item{dedik}{Dedik code}
  \item{forkod}{Parish code}
}}
\source{
Based on parish borders from Riksarkivet.
}
\usage{
data(hist_parish)
}
\description{
Swedish parish borders 1600-1990.
}

