% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{map_desc}
\alias{map_desc}
\title{Description of units}
\format{
A \code{SpatialPolygonsDataFrame} with 28 elements
\describe{
  \item{type_id}{Unit type ID}
  \item{units}{Number of units of that type}
  \item{bounds}{Number of unique years with boundary changes}
  \item{start}{Start year}
  \item{end}{End year}
}
}
\usage{
data(map_desc)
}
\description{
Description of units
}
