% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{par_to_county}
\alias{par_to_county}
\title{Parish to county}
\format{A \code{data.frame} with 2563 rows and 4 columns
\describe{
  \item{nadkod}{Parish code}
  \item{from}{Start year}
  \item{tom}{End year}
  \item{county}{County code}
}}
\source{
NAD Topography data from Riksarkivet \url{http://riksarkivet.se/psidata}.
}
\usage{
data(par_to_county)
}
\description{
Parish to county lookuptable.
}
