% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{parish_relations}
\alias{parish_relations}
\title{Parish geographical relatives}
\format{A \code{data.frame} with 2563 rows and 4 columns
\describe{
  \item{nadkod}{Parish code}
  \item{nadkod2}{Parish relative}
  \item{year}{Year of change}
  \item{relation}{Type of relation "pre" or "succ"}
}}
\usage{
data(parish_relations)
}
\description{
Parish relations, succeeding and preceding parish.
}

