% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{sp_to_ggplot}
\alias{sp_to_ggplot}
\title{Create ggplot map data.frame from NAD sp}
\usage{
sp_to_ggplot(sp)
}
\arguments{
\item{sp}{a SparialPolygonsDataFrame object}
}
\description{
Create ggplot map data.frame from NAD sp
}

