% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sweden}
\alias{sweden}
\title{Sweden}
\format{A \code{SpatialPolygonsDataFrame} with 28 elements
\describe{
  \item{SVE_ID}{id}
}}
\usage{
data(sweden)
}
\description{
Sweden national boundaries
}
