% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_sp.R
\name{clip_spdf}
\alias{clip_spdf}
\title{Clip SpatialPolygonsDataFrame}
\usage{
clip_spdf(shp, bb, byid = T)
}
\arguments{
\item{shp}{SpatialPolygonsDataFrame}

\item{bb}{Boundry box}

\item{byid}{boolean}
}
\description{
Cut a \code{SpatialPolygonsDataFrame} to a boundry box
}
