% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_boundaries.R
\name{get_boundaries}
\alias{get_boundaries}
\title{Get boundaries}
\usage{
get_boundaries(
  date,
  type = c("magistrate", "bailiwick", "hundred", "court", "municipal", "contract",
    "parish", "county", "pastorship", "diocese", "district_couryt"),
  format = "sf"
)
}
\arguments{
\item{date}{a date}

\item{type}{type of unit}

\item{format}{format of return object}
}
\description{
Get administrative boundaries
}
