% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_boundaries.R
\name{hist_boundaries}
\alias{hist_boundaries}
\title{Historical boundaries}
\usage{
hist_boundaries(date, type = c("parish", "county", "town"), format = c("sp",
  "df", "meta"))
}
\arguments{
\item{date}{a date, a year or a vector with date/year range}

\item{type}{type of unit, "parish", "county" or "town"}

\item{format}{format of return object, "df" for data.frame, "sp" for 
SpatialPolygonsDataFrame and "meta" for only meta-data.}
}
\description{
Get Swedish administrative boundaries of parishes or counties 
  for a specified year 1634-1990.
}
\examples{
map <- hist_boundaries(1900, "county")
library(sp)
plot(map)

period_map <- hist_boundaries(c(1800, 1900))
plot(period_map$map)

}
