% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{hist_town}
\alias{hist_town}
\title{County towns}
\format{A \code{SpatialPolygonsDataFrame} with 28 elements
\describe{
  \item{code}{County code}
  \item{town}{County town name}
  \item{from}{From year}
  \item{tom}{To year}
}}
\usage{
data(hist_town)
}
\description{
County towns, residensstad
}

